+++ app/src/main/java/com/app/arcabyolimpo/presentation/ui/components/atoms/buttons/DeleteAllButton.kt
@@ -14,11 +14,10 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import com.app.arcabyolimpo.presentation.theme.Poppins
-import com.app.arcabyolimpo.ui.theme.ButtonBlue
-import com.app.arcabyolimpo.ui.theme.White
-
+import com.app.arcabyolimpo.ui.theme.DangerGray
+import com.app.arcabyolimpo.ui.theme.PrimaryBlue
 /**
- * ModifyButton: blue squared button with rounded corners used to modify in the app.
+ * DeleteAllButton: blue squared button with rounded corners used to save changes in the app.
  *
  * @param onClick: () -> Unit -> function to execute when the button is clicked
  * @param cornerRadius: Dp = 8.dp -> how much rounded the corners are
@@ -27,16 +26,16 @@ import com.app.arcabyolimpo.ui.theme.White
  */
 
 @Composable
-fun ConfirmButton(
+fun DeleteAllButton(
     onClick: () -> Unit,
     cornerRadius: Dp = 8.dp,
-    width: Dp = 112.dp,
+    width: Dp = 124.dp,
     height: Dp = 40.dp,
 ) {
     Button(
         onClick = onClick,
         modifier = Modifier.size(width = width, height = height),
-        colors = ButtonDefaults.buttonColors(containerColor = ButtonBlue),
+        colors = ButtonDefaults.buttonColors(containerColor = DangerGray),
         shape = RoundedCornerShape(cornerRadius),
         contentPadding = PaddingValues(
             horizontal = 16.dp,
@@ -44,8 +43,8 @@ fun ConfirmButton(
         ),
     ) {
         Text(
-            text = "Confirmar",
-            color = White,
+            text = "Borrar Todo",
+            color = PrimaryBlue,
             fontFamily = Poppins,
             fontWeight = FontWeight.SemiBold,
         )
@@ -56,9 +55,9 @@ fun ConfirmButton(
     showBackground = true,
 )
 @Composable
-fun ConfirmButtonPreview() {
+fun DeleteAllButtonPreview() {
     MaterialTheme {
-        ConfirmButton(
+        DeleteAllButton(
             onClick = { },
         )
     }