+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## [0.1.0] - 3/11/2025
+
+### Added
+
+ - Added atom for apply button 
+ - Added atom for delete all button
+
 ## [0.1.0] - 9/10/2025
 
 ### Added
@@ -12,6 +19,7 @@
  - Compose function SaveButton created (24/10/2025)
  - Typography file in presentation/theme created to use the poppins (24/10/2025)
  - Atomic composable icons for UI (23/10/2025)
+ - Atomic composable status for UI (27/10/2025)
 
 ### Change
 - Modify constructor parameters of savebutton